Refactor `TypeExpression.DiagnosedAnalysis` to `Expression.TypeEvaluation`
