+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/ClassDefinition.kt
@@ -66,7 +66,7 @@ object ClassDefinition {
 
                 val bodyTypeConstructorAnalysis = bodyTypeConstructor.evaluateAsType()
 
-                val fieldsType = bodyTypeConstructorAnalysis.type as UnorderedTupleType
+                val fieldsType = bodyTypeConstructorAnalysis.evaluatedType as UnorderedTupleType
 
                 val tag = qualifiedPath
